Fold unlisted-realm-path endpoint tests into claim-boxel-domain-test
The standalone unlisted-realm-path-test.ts was uniquely loaded as native
ESM under the CI Node version, crashing the whole realm-server suite on a
named-import of a CJS dependency. Sibling test files with identical import
shapes load fine, and the crash did not reproduce locally. Moving the six
endpoint tests into an existing file that loads cleanly keeps the coverage
and unblocks the suite.
